type_b/app/requirements.txt
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-groq==0.3.2
langchain-huggingface==0.2.0
langchain-ollama==0.3.2
langchain-pymupdf4llm==0.4.0
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic>=2.8.4
asyncpg==0.29.0
python-dotenv==1.0.0
minio==7.1.10
psycopg==3.1.10